% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail.lines.R
\name{avail.lines}
\alias{avail.lines}
\title{Create random, available lines}
\usage{
avail.lines(poly, n, length, dist)
}
\arguments{
\item{poly}{SpatialPolygonsDataFrame, the area within which random, available lines are generated.}

\item{n}{Scalar, the number of lines to create.}

\item{length}{Scalar, the length of each line.  Unit is meters for projected coordinate systems.}

\item{dist}{Scalar, the minimum distance output lines can be from each other.  Unit is meters for projected coordinate systems.}
}
\value{
A SpatialLines object of random, available (i.e., pseudo-absence) lines.
}
\description{
This function creates random, available (i,e, pseudo-absence) lines that are within a specified polygon.
The created lines conform to a user-specified proximity rule, i.e., lines are not created within a specified
minimum distance from each other.  The direction of each line is random.  All lines have the same length.
Each line must be fully contained within the polygon.  Currently works best when input polygon is one feature.
}
\author{
Jason D. Carlisle, University of Wyoming
}

