% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail.pts.R
\name{avail.pts}
\alias{avail.pts}
\title{Create random, available points for RSF/SDM}
\usage{
avail.pts(used, poly, dist = NULL, mult = 1)
}
\arguments{
\item{used}{SpatialPointsDataFrame, the used (i.e., presence) points.}

\item{poly}{SpatialPolygonsDataFrame, the area within which random, available points are generated.}

\item{dist}{Scalar, passed to jdcR::remove.near.  The minimum distance features in pts can be from each other.  Unit is meters for projected coordinate systems.}

\item{mult}{Scalar, the ratio of available points to used points.  1 for a balanced design, 2 for twice as many available points.}
}
\value{
A SpatialPointsDataFrame of random, available (i.e., pseudo-absence) points.
}
\description{
This function creates random, available (i,e, pseudo-absence) points that are within a specified polygon.
The created points conform to a user-specified proximity rule, i.e., points are not created within a specified
minimum distance from each other or the original used points.
}
\author{
Shannon E. Albeke and Jason D. Carlisle, University of Wyoming
}

