% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fun.bootstrap.ci.R
\name{boot.mean}
\alias{boot.mean}
\title{Bootstrap CI of the mean}
\usage{
boot.mean(vec, b, ci = 0.95)
}
\arguments{
\item{vec}{Vector, must be numeric.}

\item{b}{Scalar, the number of bootstrap iterations to perform.}

\item{ci}{Scalar, the level of confidence interval (CI),  Default is 95\% CI.}
}
\value{
A named numeric vector giving the lower and upper confidence limits.
}
\description{
This function accepts a vector, removes NAs, and returns a bootstrap-generated CI of the mean.
}
\examples{
set.seed(333)
x <- runif(50, min=0, max=10)
boot.mean(vec=x, b=100, ci=0.95)
}
\author{
Jason D. Carlisle, University of Wyoming, <jason.d.carlisle@gmail.com>
}

