% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{ecoinfo}
\alias{ecoinfo}
\alias{ecoinfo-package}
\title{ecoinfo:  Assorted Tools for the Management and Analysis of Ecological Information}
\description{
Assorted tools to facilitate the management and analysis of ecological information.
Current functions focus on tools for spatial analysis and statistics.
}
\details{
Bug reports or suggestions for the ecoinfo package should be directed to <jason.d.carlisle@gmail.com>
}
