% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{jdcR}
\alias{jdcR}
\alias{jdcR-package}
\title{jdcR:  Assorted Tools for Quantitative Ecology}
\description{
This package is where I develop and maintain assorted tools that I find useful in my ecological research.  Shannon E. Albeke
(Wyoming Geographic Information Science Center, University of Wyoming) contributed to several of these tools, and each function
has contributing authors listed in its help documentation.
}
\details{
Bug reports or suggestions for the jdcR package should be directed to <jason.d.carlisle@gmail.com>
}
\author{
Jason D. Carlisle, University of Wyoming, <jason.d.carlisle@gmail.com>
}

