% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove.near.R
\name{remove.near}
\alias{remove.near}
\title{Remove spatial points based on proximity}
\usage{
remove.near(pts, dist)
}
\arguments{
\item{pts}{SpatialPointsDataFrame, the input points to thin based on proximity.}

\item{dist}{Scalar, the minimum distance features in pts can be from each other.}
}
\value{
A SpatialPointsDataFrame with equal or fewer features than pts.
}
\description{
Removes points from pts that do not meet a user-specified proximity rule, i.e., tosses
out points that are within some specified distance of another point in pts.
Note, this function is not necessarily optimized, meaning it may be possible to remove
points in a different order that would result in fewer points having to be removed.
}
\author{
Shannon E. Albeke and Jason D. Carlisle, University of Wyoming, <jason.d.carlisle@gmail.com>
}

