% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{ecoinfo}
\alias{ecoinfo}
\alias{ecoinfo-package}
\title{ecoinfo:  Assorted Tools for Ecoinformatics}
\description{
This package is a collection of tools for ecological research.  Each function
has contributing authors listed in its help documentation.
}
\details{
Bug reports or suggestions for the ecoinfo package should be directed to <jason.d.carlisle@gmail.com>
}

